Move webpack-cli out of dependencies
Webpack is not required as a dependency, only as a build step. Keep it in dev dependencies.

I'm building a phaser app with [parceljs](https://parceljs.org/) and it's working fine, but I get an annoying warning about one of the peer dependencies of `webpack-cli`: 
```
warning "phaser3-nineslice > webpack-cli@3.3.10" has unmet peer dependency "webpack@4.x.x".
```

Moving the `webpack-cli` to `devDependencies` would fix this.